pyproject.toml
[tool.poetry]
name = "rag-service"
version = "0.1.0"
description = "Production-grade multi-modal RAG system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = {extras = ["standard"], version = "^0.115.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-multipart = "^0.0.20"
aiofiles = "^24.1.0"

# Document processing
unstructured = {extras = ["pdf"], version = "^0.16.0"}
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
python-docx = "^1.1.2"
pillow = "^11.0.0"

# Audio/Video processing
openai-whisper = "^20250625"
ffmpeg-python = "^0.2.0"

# OpenAI
openai = "^2.7.0"

# Image captioning (keeping for multi-modal)
transformers = "^4.57.0"
torch = "^2.5.0"

# Vector store
qdrant-client = "^1.13.0"

# BM25
rank-bm25 = "^0.2.2"

# Task queue
redis = "^5.2.0"
rq = "^2.0.0"

# Utilities
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
httpx = "^0.28.0"
tiktoken = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.8.0"
ipython = "^8.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
